sounds=true, because so many complaints on 0.8.4
sounds=false by default was not the intended behaviour, as I recall. Users have not taken it well, either. Add to that the fact that you have to start the client, turn on sounds, and then restart the client, and... I'm just changing that back to true, before the next release.