Added support for the Real Time Bridge command (RTB), to allow notification of new notes and tracking bookmark and friends changes. RTB: 'note' message will display a notice with a hyperlink to the note in question. RTB: 'trackadd' and 'friendadd' will display a message the character has been added too bookmarks or as a friend respectively. This will also update the stored list of friends. RTB: 'trackrem' and 'friendremove' will display a message the character is no longer bookmarked or is no longer a friend respectively. Currently this does not update the stored list of friends, as there it's not certain that all instances bookmarks and friends have been removed for that character. RTB: 'friendrequest' will display a message that another character has sent a friend request, and will also display a link to the relevant page.
Added 'friendupdate.wav' for use with friend updates in the RTB. It is currently a copy of 'newnote.wav'.
Added SOUND_FRIENDUPDATE the enum in FSound.
New notes and friend updates should play sounds. Bookmark updates are silent.

Added a GENDER_OFFLINE_UNKNOWN to the gender enum in FCharacter. This is to handle cases where a character name should be hyperlinked and colored, but they're offline for whatever reason.

Added FSession.getCharacterUrl() to get the URL of a character that is offline, or not checked for their online/offline state.
Added FSession.getCharacterHtml() to generate a properly formatted HTML link of a character, and colored correctly if they're online.